+++ scopes/react/component-highlighter/children-highlighter/index.ts
@@ -1,0 +1,2 @@
+export { MultiHighlighter } from './children-highlighter';
+export type { MultiHighlighterProps } from './children-highlighter';
