+++ scopes/react/component-highlighter/children-highlighter/use-children-highlighter.tsx
