+++ scopes/react/component-highlighter/hybrid-highlighter/hybrid-highlighter.tsx
@@ -10,7 +10,7 @@ import {
   HighlightClasses,
   HighlighterSize,
 } from '../element-highlighter';
-import { useMultiHighlighter } from '../multi-highlighter/use-multi-highlighter';
+import { useMultiHighlighter } from '../children-highlighter/use-children-highlighter';
 import type { MatchRule, ComponentMatchRule } from '../rule-matcher';
 
 export interface HybridHighlighterProps extends React.HTMLAttributes<HTMLDivElement> {
