+++ scopes/react/component-highlighter/hybrid-highlighter/index.ts
