--- scopes/react/component-highlighter/multi-highlighter/index.ts
@@ -1,2 +1,0 @@
-export { MultiHighlighter } from './multi-highlighter';
-export type { MultiHighlighterProps } from './multi-highlighter';
